tracking/urls.py
from django.urls import path
from . import views

from django.urls import path
from rest_framework_simplejwt.views import ( # type: ignore
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('next-tracking-number/', views.NextTrackingNumberView.as_view(), name='next_tracking_number'),
]

